Fix bug with lasso settings

+++ scripts/ml/lasso.py
@@ -22,9 +22,9 @@ class Lasso(Mlinterface):
             self.n_split_shuffle(samples_with_names, target, int(self.config["n"]))
 
         if self.config["positive"].lower() == "true":
-            clf = linear_model.Lasso(positive=True, tol=int(self.config["tol"]))
+            clf = linear_model.Lasso(positive=True, tol=float(self.config["tol"]))
         else:
-            clf = linear_model.Lasso(positive=False, tol=int(self.config["tol"]))
+            clf = linear_model.Lasso(positive=False, tol=float(self.config["tol"]))
         score, predictions = self.make_predictions(clf, train_sample, train_target, test_sample, test_target, test_name)
 
         auc_scores = []
